Extract the world manager code into core
