+++ src/components/Loader.vue
@@ -12,9 +12,7 @@
 import { mapState } from 'vuex'
 
 export default {
-  computed: {
-    ...mapState(['config']),
-  },
+  computed: { ...mapState(['config']) },
 }
 </script>
