+++ src/core/Cell.js
