+++ src/core/worldManager.js
