+++ src/store.js
@@ -2,11 +2,11 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 
 import experimentManager from './experiments'
-import worldManager from './worldManager'
+import worldManager from './core/worldManager'
 
 Vue.use(Vuex)
 
-export default new Vuex.Store({
+const store = new Vuex.Store({
   state: {
     config: { columns: 100, rows: 50, cellSize: 7 },
     generation: 0,
@@ -15,24 +15,21 @@ export default new Vuex.Store({
     loading: { experiment: true, renderer: true },
     fps: 60,
     experiments: experimentManager.experiments,
-    currentExperiment: experimentManager.defaultExperiment,
+    currentExperiment: experimentManager.defaultExperimentName,
     renderer: 'Canvas',
   },
 
   actions: {
     async init({ state, commit }, config = state.config) {
       config !== state.config && commit('changeConfig', config)
-
-      worldManager.setExperiments(experimentManager.experiments)
-      worldManager.on('init', () => commit('setLoading', { experiment: false }))
-      worldManager.on('update', nextGeneration => commit('nextGeneration', nextGeneration))
-
       await worldManager.init(config)
     },
 
     update() { worldManager.update() },
-
     restart() { worldManager.restart() },
+    setLoading({ commit }, loading) { commit('setLoading', loading) },
+    play({ commit }) { commit('setIsPlaying', worldManager.play()) },
+    pause({ commit }) { commit('setIsPlaying', worldManager.pause()) },
 
     changeExperiment({ state, commit }, experiment) {
       if (state.currentExperiment !== experiment) {
@@ -50,20 +47,6 @@ export default new Vuex.Store({
     changeConfig({ commit }, partialConfig) {
       commit('changeConfig', partialConfig)
       worldManager.setConfig(partialConfig)
-    },
-
-    setLoading({ commit }, loading) {
-      commit('setLoading', loading)
-    },
-
-    play({ commit }) {
-      commit('setIsPlaying', true)
-      worldManager.play()
-    },
-
-    pause({ commit }) {
-      commit('setIsPlaying', false)
-      worldManager.pause()
     },
 
     changeFPS({ commit }, fps) {
@@ -104,3 +87,9 @@ export default new Vuex.Store({
     },
   },
 })
+
+worldManager.setExperiments(experimentManager.experiments)
+worldManager.on('init', () => store.commit('setLoading', { experiment: false }))
+worldManager.on('update', nextGeneration => store.commit('nextGeneration', nextGeneration))
+
+export default store
